Move type definitions to their own file
